% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplR.R
\docType{package}
\name{eplR}
\alias{eplR}
\title{\code{eplR} package}
\description{
The goal of eplR is to make it easy for any user of R to access and analyze data from the
English Premier League (soccer). The package provides R functions that simplify and expedite
common tasks for studying match, referee and betting statistics in the league.
}
\details{
See the README on
\href{https://github.com/gl2668/eplR/blob/master/README.md}{GitHub}
}
