% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referee.R
\name{referee}
\alias{referee}
\title{referee}
\usage{
referee(season)
}
\arguments{
\item{season}{The season argument allows the user to specify which season he or she wants to get EPL data for.
Ensure that season is in the correct character string format, e.g. "1819" for 2018 - 2019 Season
and "1718" for 2017 - 2018 season.}
}
\value{
A dataframe with season statistics for the specified Season.
}
\description{
This function allows R users to get referee statistics for a specified season in the
English Premier League. The data is retrieved from DataHub.io and football-data.co.uk.
It covers from 2002 to 2019.
}
\details{
The function requires users to input one argument. This argument specifies the EPL Season
that the user wants to retrieve data for.
}
\examples{
referee(season = "1819")
referee("1617")
}
