% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamData.R
\name{teamData}
\alias{teamData}
\title{teamData}
\usage{
teamData(season, team)
}
\arguments{
\item{season}{The season argument allows the user to specify for which season he or she wants to get EPL data.
Ensure that season is in the correct character string format, e.g. "1819" for 2018 - 2019 Season
and "1718" for 2017 - 2018 season.}

\item{team}{The team argument requires the user to specify the team he or she wants the data for. The first letter
of the team name should be capitalized or if the team is name is 2 words, e.g. Man United - you should
capitalize the first letter of both words.}
}
\value{
A dataframe with summary season statistics for the specified team and season.
}
\description{
This function allows R users to get season statistics for a specified team and
season in the English Premier League. The data is retrieved from DataHub.io and
football-data.co.uk. It covers from 2009 to 2019.
}
\details{
The function requires users to input two different arguments. The first argument
specifies the EPL Season that he or she wants to retrieve data for. The second argument
specifies which team he or she wants to get data for.
}
\examples{
teamData(season = "1819", team = "Liverpool")
teamData("1617", "Man United")
}
