% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamPerformance.R
\name{teamPerformance}
\alias{teamPerformance}
\title{teamPerformance}
\usage{
teamPerformance(season, team)
}
\arguments{
\item{season}{The season argument allows the user to specify for which season he or she wants to get EPL
e.g. "1819" for 2018 - 2019 Season; and "1718" for 2017 - 2018 season.}

\item{team}{The team argument allows the user to filter the data to visualize the season performance of the team
that the user is only interested in. Ensure that team name is in correct character string format, e.g. "Chelsea" or "Man City".}
}
\value{
A plot of the specified team's cummulated points over the specified Season.
}
\description{
This function allows R users to visualize the performance of a specified team in the English Premier League
 over a specified season.
}
\details{
The function requires the user to input two arguments. The first argument is the season and the second
argument is the team.
}
\examples{
teamPerformance(season = "1819", team = "Arsenal")
teamPerformance(season = "1617", "Chelsea")
teamPerformance(1112, "Wolves")
}
